ability to add new namespace and create workflow in a namespace

+++ api/src/main/java/com/cognitree/kronos/api/WorkflowDefinitionResource.java
@@ -18,10 +18,10 @@
 package com.cognitree.kronos.api;
 
 import com.cognitree.kronos.model.Workflow;
-import com.cognitree.kronos.model.WorkflowId;
 import com.cognitree.kronos.model.definitions.WorkflowDefinition;
 import com.cognitree.kronos.model.definitions.WorkflowDefinitionId;
 import com.cognitree.kronos.scheduler.WorkflowSchedulerService;
+import com.cognitree.kronos.scheduler.store.NamespaceStoreService;
 import com.cognitree.kronos.scheduler.store.WorkflowDefinitionStoreService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -33,6 +33,7 @@ import org.slf4j.LoggerFactory;
 
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
+import javax.ws.rs.HeaderParam;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
@@ -47,19 +48,21 @@ import static javax.ws.rs.core.Response.Status.CONFLICT;
 import static javax.ws.rs.core.Response.Status.NOT_FOUND;
 import static javax.ws.rs.core.Response.Status.OK;
 
-@Path("workflowdefs")
-@Api(value = "workflow definition", description = "manage workflow definitions")
+@Path("/definitions/{res:workflows|w}")
+@Api(value = "workflow definitions", description = "manage workflow definitions")
 public class WorkflowDefinitionResource {
     private static final Logger logger = LoggerFactory.getLogger(WorkflowDefinitionResource.class);
-
-    private static final String DEFAULT_NAMESPACE = "default";
 
     @GET
     @ApiOperation(value = "Get all workflow definitions", response = WorkflowDefinition.class, responseContainer = "List")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getAllWorkflowDefinitions() {
-        logger.info("Received request to get all workflow definitions");
-        final List<WorkflowDefinition> workflowDefinitions = WorkflowDefinitionStoreService.getService().load();
+    public Response getAllWorkflowDefinitions(@HeaderParam("namespace") String namespace) {
+        logger.info("Received request to get all workflow definitions under namespace {}", namespace);
+        if (!validateNamespace(namespace)) {
+            return Response.status(BAD_REQUEST).entity("no namespace exists with name " + namespace).build();
+        }
+        final List<WorkflowDefinition> workflowDefinitions =
+                WorkflowDefinitionStoreService.getService().load(namespace);
         return Response.status(OK).entity(workflowDefinitions).build();
     }
 
@@ -70,13 +73,18 @@ public class WorkflowDefinitionResource {
             @ApiResponse(code = 404, message = "Workflow definition not found")})
     @Produces(MediaType.APPLICATION_JSON)
     public Response getWorkflowDefinition(@ApiParam(value = "workflow definition name", required = true)
-                                          @PathParam("name") String workflowDefName) {
-        logger.info("Received request to get workflow definition with name {}", workflowDefName);
-        WorkflowDefinitionId workflowDefinitionId = WorkflowDefinitionId.create(workflowDefName, DEFAULT_NAMESPACE);
+                                          @PathParam("name") String workflowDefName,
+                                          @HeaderParam("namespace") String namespace) {
+        logger.info("Received request to get workflow definition with name {} under namespace {}",
+                workflowDefName, namespace);
+        if (!validateNamespace(namespace)) {
+            return Response.status(BAD_REQUEST).entity("no namespace exists with name " + namespace).build();
+        }
+        WorkflowDefinitionId workflowDefinitionId = WorkflowDefinitionId.create(workflowDefName, namespace);
         final WorkflowDefinition workflowDefinition =
                 WorkflowDefinitionStoreService.getService().load(workflowDefinitionId);
         if (workflowDefinition == null) {
-            logger.error("No workflow definition exists with name {}", workflowDefName);
+            logger.error("No workflow definition exists with name {} under namespace {}", workflowDefName, namespace);
             return Response.status(NOT_FOUND).build();
         }
         return Response.status(OK).entity(workflowDefinition).build();
@@ -87,12 +95,17 @@ public class WorkflowDefinitionResource {
     @ApiResponses(value = {
             @ApiResponse(code = 409, message = "Workflow definition already exists")})
     @Produces(MediaType.APPLICATION_JSON)
-    public Response addWorkflowDefinition(WorkflowDefinition workflowDefinition) {
-        logger.info("Received request to add workflow definition {}", workflowDefinition);
-        // override namespace of workflow
-        workflowDefinition.setNamespace(DEFAULT_NAMESPACE);
+    public Response addWorkflowDefinition(@HeaderParam("namespace") String namespace,
+                                          WorkflowDefinition workflowDefinition) {
+        logger.info("Received request to add workflow definition {} under namespace {}",
+                workflowDefinition, namespace);
+        if (!validateNamespace(namespace)) {
+            return Response.status(BAD_REQUEST).entity("no namespace exists with name " + namespace).build();
+        }
+        workflowDefinition.setNamespace(namespace);
         if (WorkflowDefinitionStoreService.getService().load(workflowDefinition) != null) {
-            logger.error("Workflow definition already exists with name {}", workflowDefinition.getName());
+            logger.error("Workflow definition already exists with name {} under namespace {}",
+                    workflowDefinition.getName(), namespace);
             return Response.status(CONFLICT).build();
         }
         try {
@@ -111,13 +124,18 @@ public class WorkflowDefinitionResource {
             @ApiResponse(code = 404, message = "Workflow definition not found")})
     @Produces(MediaType.APPLICATION_JSON)
     public Response executeWorkflowDefinition(@ApiParam(value = "workflow definition name", required = true)
-                                              @PathParam("name") String workflowDefName) {
-        logger.info("Received request to execute workflow definition with name {}", workflowDefName);
-        WorkflowDefinitionId workflowDefinitionId = WorkflowDefinitionId.create(workflowDefName, DEFAULT_NAMESPACE);
+                                              @PathParam("name") String workflowDefName,
+                                              @HeaderParam("namespace") String namespace) {
+        logger.info("Received request to execute workflow definition with name {} under namespace {}",
+                workflowDefName, namespace);
+        if (!validateNamespace(namespace)) {
+            return Response.status(BAD_REQUEST).entity("no namespace exists with name " + namespace).build();
+        }
+        WorkflowDefinitionId workflowDefinitionId = WorkflowDefinitionId.create(workflowDefName, namespace);
         final WorkflowDefinition workflowDefinition =
                 WorkflowDefinitionStoreService.getService().load(workflowDefinitionId);
         if (workflowDefinition == null) {
-            logger.error("No workflow definition exists with name {}", workflowDefName);
+            logger.error("No workflow definition exists with name {} under namespace {}", workflowDefName, namespace);
             return Response.status(NOT_FOUND).build();
         }
         Workflow workflow = WorkflowSchedulerService.getService()
@@ -133,13 +151,17 @@ public class WorkflowDefinitionResource {
     @Produces(MediaType.APPLICATION_JSON)
     public Response updateWorkflowDefinition(@ApiParam(value = "workflow definition name", required = true)
                                              @PathParam("name") String workflowDefName,
+                                             @HeaderParam("namespace") String namespace,
                                              WorkflowDefinition workflowDefinition) {
-        logger.info("Received request to update workflow definition with name {} to {}",
-                workflowDefName, workflowDefinition);
+        logger.info("Received request to update workflow definition with name {} under namespace {} to {}",
+                workflowDefName, namespace, workflowDefinition);
+        if (!validateNamespace(namespace)) {
+            return Response.status(BAD_REQUEST).entity("no namespace exists with name " + namespace).build();
+        }
         workflowDefinition.setName(workflowDefName);
-        workflowDefinition.setNamespace(DEFAULT_NAMESPACE);
+        workflowDefinition.setNamespace(namespace);
         if (WorkflowDefinitionStoreService.getService().load(workflowDefinition) == null) {
-            logger.error("No workflow definition exists with name {}", workflowDefName);
+            logger.error("No workflow definition exists with name {} under namespace {}", workflowDefName, namespace);
             return Response.status(NOT_FOUND).build();
         }
         try {
@@ -156,17 +178,25 @@ public class WorkflowDefinitionResource {
     @ApiOperation(value = "Delete workflow definition")
     @ApiResponses(value = {
             @ApiResponse(code = 404, message = "Workflow definition not found")})
-    @Produces(MediaType.APPLICATION_JSON)
     public Response deleteWorkflowDefinition(@ApiParam(value = "workflow definition name", required = true)
-                                             @PathParam("name") String workflowDefName) {
-        logger.info("Received request to delete workflow definition with name {}", workflowDefName);
-        WorkflowDefinitionId workflowDefinitionId = WorkflowDefinitionId.create(workflowDefName, DEFAULT_NAMESPACE);
+                                             @PathParam("name") String workflowDefName,
+                                             @HeaderParam("namespace") String namespace) {
+        logger.info("Received request to delete workflow definition with name {} under namespace {}",
+                workflowDefName, namespace);
+        if (!validateNamespace(namespace)) {
+            return Response.status(BAD_REQUEST).entity("no namespace exists with name " + namespace).build();
+        }
+        WorkflowDefinitionId workflowDefinitionId = WorkflowDefinitionId.create(workflowDefName, namespace);
         if (WorkflowDefinitionStoreService.getService().load(workflowDefinitionId) == null) {
-            logger.error("No workflow definition exists with name {}", workflowDefName);
+            logger.error("No workflow definition exists with name {} under namespace {}", workflowDefName, namespace);
             return Response.status(NOT_FOUND).build();
         }
         WorkflowSchedulerService.getService().delete(workflowDefinitionId);
         WorkflowDefinitionStoreService.getService().delete(workflowDefinitionId);
         return Response.status(OK).build();
+    }
+
+    private boolean validateNamespace(String name) {
+        return name != null && NamespaceStoreService.getService().load(name) != null;
     }
 }
